refactor: move core objects to a separate assembly
this will provide a CLI command system and repository representation to be used in plugins
