+++ cmf-cli.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tests", "tests\tests.csproj", "{C9C222C5-6236-4991-B3CF-6289E2ACBF09}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "core", "core\core.csproj", "{BF10AAE5-ACB7-47FD-9C85-F0AAEF1A7820}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,6 +28,10 @@ Global
 		{C9C222C5-6236-4991-B3CF-6289E2ACBF09}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9C222C5-6236-4991-B3CF-6289E2ACBF09}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9C222C5-6236-4991-B3CF-6289E2ACBF09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF10AAE5-ACB7-47FD-9C85-F0AAEF1A7820}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF10AAE5-ACB7-47FD-9C85-F0AAEF1A7820}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF10AAE5-ACB7-47FD-9C85-F0AAEF1A7820}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF10AAE5-ACB7-47FD-9C85-F0AAEF1A7820}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
