+++ core/Attributes/CmfCommandAttribute.cs
