+++ core/Enums/ContentType.cs
