+++ core/Enums/MessageType.cs
