+++ core/Objects/ContentToPack.cs
