+++ core/Objects/DependencyCollection.cs
