+++ core/Objects/FileToPack.cs
