+++ core/Objects/RepositoriesConfig.cs
