+++ core/Objects/VersionService.cs
@@ -1,0 +1,13 @@
+using System.Reflection;
+
+namespace Cmf.Common.Cli.Objects
+{
+    /// <summary>
+    /// Interface for a service that returns the current (running) CLI version
+    /// </summary>
+    public interface IVersionService
+    {
+        public string PackageId { get; }
+        string CurrentVersion { get;  }
+    }
+}
