+++ core/Utilities/CliException.cs
