+++ package.json
@@ -30,10 +30,10 @@
     "bump:npm:feature": "cd npm && npm version minor --no-git-tag-version && cd ..",
     "bump:npm:breaking": "cd npm && npm version major --no-git-tag-version && cd ..",
     "bump:npm:dev": "cd npm && npm version $DEV_VERSION --no-git-tag-version && cd ..",
-    "bump:dotnet:pre": "cd cmf-cli && dotnet-bump prerelease --force --no-commit && cd ..",
-    "bump:dotnet:patch": "cd cmf-cli && dotnet-bump patch --force --no-commit && cd ..",
-    "bump:dotnet:feature": "cd cmf-cli && dotnet-bump minor --force --no-commit && cd ..",
-    "bump:dotnet:breaking": "cd cmf-cli && dotnet-bump major --force --no-commit && cd ..",
+    "bump:dotnet:pre": "cd cmf-cli && dotnet-bump prerelease --force --no-commit && cd .. && cd core && dotnet-bump prerelease --force --no-commit && cd ..",
+    "bump:dotnet:patch": "cd cmf-cli && dotnet-bump patch --force --no-commit && cd .. && cd core && dotnet-bump patch --force --no-commit && cd ..",
+    "bump:dotnet:feature": "cd cmf-cli && dotnet-bump minor --force --no-commit && cd .. && cd core && dotnet-bump minor --force --no-commit && cd ..",
+    "bump:dotnet:breaking": "cd cmf-cli && dotnet-bump major --force --no-commit && cd .. && cd core && dotnet-bump major --force --no-commit && cd ..",
     "gen:changelog": "cd npm && standard-version && cd ..",
     "commit": "git-cz"
   },
